# Speedup iteration for wrappers around arrays