This will make `for` loops such as `for i in itr; #= body =#; end` lower to

```julia
@inbounds next = iterate(iter)
while next !== nothing
    (i, state) = next
    # body
    @inbounds next = iterate(iter, state)
end
```

Currently no `@inbounds` annotations are added unless the whole for loop is manually annotated. This should make, for instance, iteration over `views` faster.

 Before
```julia
using BenchmarkTools

let
    v = rand(100);
    w = view(v, :)
    @btime any(iszero, $v)  # 41.471 ns (0 allocations: 0 bytes)
    @btime any(iszero, $w)  # 61.306 ns (0 allocations: 0 bytes)
end
```

After
```julia
using BenchmarkTools

let
    v = rand(100);
    w = view(v, :)
    @btime any(iszero, $v)  # 41.471 ns (0 allocations: 0 bytes)
    @btime any(iszero, $w)  # 41.840 ns (0 allocations: 0 bytes)
end
```

Also fixes #39354.