@JeffBezanson 2021-03-25
I believe we've considered this way too dangerous in the past. It basically makes part of the language unsafe by default.