@pabloferz 2021-03-25
Oh right. Any `for` loop would lower `iterate` but if iterate is marked incorrectly as `@inbounds` things will, break. Will have to think of something else.